Handles mult and div, but everything's a float now

+++ rpn.h
@@ -1,5 +1,5 @@
 struct Stack_value{
-	int value;
+	float value;
 	struct Stack_value* prev;
 };
 
@@ -7,5 +7,5 @@ struct Stack{
 	struct Stack_value* top;
 };
 
-int pop(struct Stack*);
-void push(struct Stack*, int);
+float pop(struct Stack*);
+void push(struct Stack*, float);
